requirements.txt
# Auto Fix Agent Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
py7zr>=0.20.0

# AWS Integration
boto3>=1.34.0

# Code Scanning & Lint
pylint>=3.0.0
flake8>=6.0.0
ruff>=0.4.0
black>=24.0.0
radon>=6.0.1

# Security
bandit>=1.7.5
semgrep>=1.45.0

# Testing
pytest>=8.0.0

# Git Integration
gitpython>=3.1.40

# LLM Integration
openai>=1.10.0
tiktoken>=0.5.2

# Utilities
rich>=13.7.0
click>=8.1.7
